+++ src/core/src/OmniIonServer.cpp
@@ -17,7 +17,9 @@ OmniIonServer::OmniIonServer(Context* pContext, const pxr::SdfPath& path)
           pContext->getHttpAssetAccessor(),
           getIonServerUrl(),
           getIonServerApiUrl(),
-          getIonServerApplicationId())) {}
+          getIonServerApplicationId())) {
+    _session->resume();
+}
 
 const pxr::SdfPath& OmniIonServer::getPath() const {
     return _path;
@@ -29,6 +31,9 @@ std::shared_ptr<CesiumIonSession> OmniIonServer::getSession() const {
 
 std::string OmniIonServer::getIonServerUrl() const {
     const auto cesiumIonServer = UsdUtil::getCesiumIonServer(_pContext->getUsdStage(), _path);
+    if (!UsdUtil::isSchemaValid(cesiumIonServer)) {
+        return "";
+    }
 
     std::string ionServerUrl;
     cesiumIonServer.GetIonServerUrlAttr().Get(&ionServerUrl);
@@ -38,6 +43,9 @@ std::string OmniIonServer::getIonServerUrl() const {
 
 std::string OmniIonServer::getIonServerApiUrl() const {
     const auto cesiumIonServer = UsdUtil::getCesiumIonServer(_pContext->getUsdStage(), _path);
+    if (!UsdUtil::isSchemaValid(cesiumIonServer)) {
+        return "";
+    }
 
     std::string ionServerApiUrl;
     cesiumIonServer.GetIonServerApiUrlAttr().Get(&ionServerApiUrl);
@@ -47,6 +55,9 @@ std::string OmniIonServer::getIonServerApiUrl() const {
 
 int64_t OmniIonServer::getIonServerApplicationId() const {
     const auto cesiumIonServer = UsdUtil::getCesiumIonServer(_pContext->getUsdStage(), _path);
+    if (!UsdUtil::isSchemaValid(cesiumIonServer)) {
+        return 0;
+    }
 
     int64_t ionServerApplicationId;
     cesiumIonServer.GetIonServerApplicationIdAttr().Get(&ionServerApplicationId);
@@ -56,6 +67,9 @@ int64_t OmniIonServer::getIonServerApplicationId() const {
 
 CesiumIonClient::Token OmniIonServer::getToken() const {
     const auto cesiumIonServer = UsdUtil::getCesiumIonServer(_pContext->getUsdStage(), _path);
+    if (!UsdUtil::isSchemaValid(cesiumIonServer)) {
+        return {};
+    }
 
     std::string projectDefaultIonAccessToken;
     std::string projectDefaultIonAccessTokenId;
@@ -72,6 +86,9 @@ CesiumIonClient::Token OmniIonServer::getToken() const {
 
 void OmniIonServer::setToken(const CesiumIonClient::Token& token) {
     const auto cesiumIonServer = UsdUtil::getCesiumIonServer(_pContext->getUsdStage(), _path);
+    if (!UsdUtil::isSchemaValid(cesiumIonServer)) {
+        return;
+    }
 
     cesiumIonServer.GetProjectDefaultIonAccessTokenAttr().Set(token.token);
     cesiumIonServer.GetProjectDefaultIonAccessTokenIdAttr().Set(token.id);